完成批量增加用户功能

+++ back/src/main/java/com/fzu/teamwork/service/BlockServiceImpl.java
@@ -3,6 +3,7 @@ package com.fzu.teamwork.service;
 import com.fzu.teamwork.dao.BlockDao;
 import com.fzu.teamwork.model.Block;
 import com.fzu.teamwork.model.User;
+import com.sun.org.apache.xpath.internal.operations.Bool;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.RestController;
@@ -18,16 +19,26 @@ public class BlockServiceImpl implements BlockService {
     @Resource
     private BlockDao blockDao;
 
-    //add block
+    //添加临时板块（返回添加结果）
     @Override
-    public void addBlock(Block block) {
-        blockDao.insert(block);
+    public Boolean addBlock(Block block) {
+        if(blockDao.selectByExample(null).size() > 0){//已经存在临时板块
+            return false;//添加失败
+        }else{
+            blockDao.insert(block);
+            return true;//添加成功
+        }
     }
 
     //删除临时板块
     @Override
-    public void deleteBlock(){
-        blockDao.deleteByExample(null);
+    public boolean deleteBlock(){
+        if(blockDao.selectByExample(null).size() == 0){//临时板块不存在
+            return false;
+        }else{
+            blockDao.deleteByExample(null);
+            return true;
+        }
     }
 
     //获取临时板块
